Add getSkillPAGES In getSkill and getSkillPAGES, computing stds by hand because using np.std was giving weird results (possibly due to the different normalization)

+++ t_subsample.py
@@ -30,7 +30,7 @@ interval: two-element array giving the start and end years. Make sure that these
 
     for ii in np.arange(0,len((recdf.columns))):
         #rec = recdf.iloc[interval[0]:(interval[1]-np.mod(intl,tau)),ii].values
-        rec = recdf.iloc[interval[0]:interval[1],ii].values
+        rec = recdf.iloc[interval[0]:interval[1],ii].values.astype('float')
         rr = rec.reshape(tau,-1)
         r2 = np.nanmean(rr,0)
         df[recdf.columns[ii]] = pandas.DataFrame(data=r2,index=newt)
